Fixed crash when setting header data in QSqlQueryModel.
The crash (Q_ASSERT_X failure) happened when a proxy model was being attached
to the QSqlQueryModel, and no query was set yet. The headerDataChanged()
signal was being received by the proxy model who wouldn't check its
"proxyfied" data bounds.

The patch introduces a behaviour change. However, this change makes the usage
of QSqlQueryModel::setHeaderData() to be more in accordance with the current
documentation, and to behave in the same way as for QStandardItemModel,
QTreeModel, and QTableModel.

Task-number: QTBUG-4963
Reviewed-by: Olivier