Cocoa: Synthesize italic text correct way when adding glyphs to path
On Mac OS X, Cocoa, we would synthesize italics on the text by slanting
it in the incorrect direction (so it leaned to the left) when generating
a path from the text, e.g. when printing.

The patch makes the text slant the correct way, and the logic now
becomes identical with the synthesized italics in the draw() function.

Task-number: QTBUG-4969
Reviewed-by: Trond