QGAL: Add test for David Boddie bug
The current simplification code does not handle sequences with anchors
pointed to different directions could not be simplified together into
a sequential anchor.

       A (10 / 20 / 50 )   B (20 / 20 / 20)
Ex:   o-----------------> <----------------o

The reason we don't support it yet is shown in the example above. The
resulting anchor would be either:

             Result_1 (-10 / 0 / 30)
      o------------------------------------>
                                                 (or)

             Result_2 (-30 / 0 / 10)
      <------------------------------------o

But the current implementation assumes no anchors can have negative
sizes.

Hopefully, the next commits will add support for it and then enable
such simplification.  :-)

Signed-off-by: Eduardo M. Fleury <eduardo.fleury@openbossa.org>