Switched setWindowIcon_sys to use QPixmpa::toSymbianCFbsBitmap in S60.
There were TODOs in code to remove the temporary solution for creating
native CFbsBitmap out of QPixmap. Now when QPixmpa has native backed
and it provides toSymbianCFbsBitmap, it it preferred to use
toSymbianCFbsBitmap since in best case it can only duplicate the bitmap
handle instead of copying the data.

Task-number: QTBUG-4948
Reviewed-by: Jani Hautakangas