Selected tab is drawn incorrectly on Snow Leopard.
The default height of tab bar is 22 pixels in Snow Leopard. We used to
draw tabs taller than 21 pixels to a pixmap and stretch to the required
size. The limit is now changed to 22 pixels (which is the most common
use case). The stretched drawing is not perfect in Snow Leopard due to
some changes in how HITheme draws tabs.

Reviewed-by: Jens Bache-Wiig