Fix the combobox popup 1-second fadeout regression on Carbon.
Regression caused by b946da648af0c5fa1c73fe1e57b0b1e08fb14d13.

Prior to that commit, the default duration for macWindowFade was 0, but there
was code in the implementation that set it to 0.15 in that case.

Set the default duration to 0.15. This matches the old behavior when calling
macWindowFade without specifying a duration, and makes it clearer what the
default really is.