Fix tst_QBoxLayout::setGeometry failure on S60 3.x
The S60 style posts a LayoutRequest event when a widget is shown, so
that it can draw the focus rectangle around the widget that has keyboard
focus.
Although lay2->setGeometry was working correctly, processEvents() caused
the outer layout to be re-laid out (which expands the inner layout and
QDial to fill the available space).
The processEvents() call is not necessary for the test case, so it can
be removed.

Reviewed-by: Jan-Arve