Do not display transient binding errors
During QML startup, it is common to have "errors" in bindings as the apps
state stabilizes.  These are not real errors, but just a consequence of
implementing a declarative UI in an imperative world.  Now during startup,
the display of errors is delayed until the startup completes, and then only
bindings that are still in an error state are displayed.

QT-2373