Let QGraphicsItem set QObjectPrivate::wasDeleted when appropriate.
This change allows children of QGraphicsObject-derived classes to check
if the parent is being deleted by checking the private flag wasDeleted.

Reverts 37b16d5cbb4e7bc534f690ebf50434d228b5ecfc, p4 change 9681, to
allow QGraphicsItem to set QObjectPrivate's wasDeleted member before
entering QObjectPrivate's destructor. The original code was in there to
let the user know, via the console output, that QObject was
double-deleted (e.g., when placing QObject on the stack, and also giving
it a parent, so that if the parent is deleted first, bang).

Reviewed-by: Aaron Kennedy
Reviewed-by: Bradley T. Hughes