Fix initial focus bug in ItemIsFocusScope.
The task provides an example that doesn't gain input focus when started.
The fix contains two parts: one is to allow items with focus scope
ancestors to become focus items even if the scope is inactive. The other
is to fix up the focusItem pointers on reparent. Before, the focus
scopes' focusItem pointers always pointed to itself, or the next scope.
Now these items are treated no differently than other items in that
respect.

The change has a performance impact when reparenting a large subtree
that has a sub focus item onto another item (one more dig).

Task-number: QT-2331
Reviewed-by: Alexis Menard