Greatly improve the performance of obtaining a state's transitions
Transitions are children of their source state. We use
QObject::children() and qobject_cast() each child to a
QAbstractTransition to see if it is indeed a transition. However,
calling qobject_cast() is very expensive. This commit introduces
a cached list of transitions. The list is invalidated after a
child object has been added or removed.

In the typical case we expect the object hierarchy to remain fairly
constant once the state machine has been started (states, child states
and transitions are usually "static"), in other words the cached list
is not likely to be invalidated much.

Obtaining a state's transitions needs to be as fast as possible
because it's in the critical path of the state machine algorithm.

Reviewed-by: Eskil Abrahamsen Blomfeldt