Fixed crash/drawing artifacts on rotation change on Symbian.
On every beginDataAccess on the pixmap data, we checked the value of
the pointer to the bits and only updated the image if the pointer had
changed. However, we didn't take into account that the pointer could
be the same, even though the dimensions were different, since
malloc() could return the same memory area. This would lead to
painting into an image that had the wrong dimensions, which again led
to either crashes or image shearing. Fixed by checking the dimensions
before deciding to update the image.

Task:     QTBUG-4815
RevBy:    Jason Barron