Modified gesture events propagation.
By default if the gesture is ignored, only gestures in the started state
are propagated, and accepting a gesture in the started state adds an
implicit grab meaning all the following events in the gesture sequence
will be delivered to that widget. This is similar to the way QTouchEvent
is propagated.

Also added a hint, which specifies if gestures in any state can be
propagated to the widget which has enabled the hint.

Reviewed-by: Thomas Zander