Fix QDateTime on S60 3.1
S60 3.1 plugin returned empty strings for the unsupported APIs
Since these APIs are needed for the most common use case of converting
a QDateTime to a QString using the local format, I have implemented an
emulation of the missing APIs using older APIs that are supported.

Updated the autotest so it does some sanity checking on the result of
local date/time conversion - it would pass instead of fail if the string
was garbage before.

Reviewed-by: Aleksandar Sasha Babic