Fix tst_QWidget::compatibilityChildInsertedEvents test
This broke after commit: a8e2a457bb7d2fc377c1c65b6a4172974919e055
(Add a new event type, WinIdChange). The first thing that happens in
show() is that the top-level is created, and hence we send WinIdChange.
We therefore have to add this event to list of expected events.

Reviewed-by: TrustMe