Cache a state's parent state
QAbstractState::parentState() is called heavily by the state machine
algorithm. The parent state is obtained by qobject_cast'ing
QObject::parent(). qobject_cast() is expensive. This commit introduces
caching of the result in order to improve performance.

We expect that the cache won't be invalidated much since the parent-child
relationship of states usually doesn't change after the state machine is
started.

Reviewed-by: Eskil Abrahamsen Blomfeldt