Introduce internal StateType to avoid excessive qobject_casts
The state machine algorithm frequently needs to know what type a state
is, e.g. if it is atomic, final or a history state. We were using
qobject_cast() to determine this, but that function is expensive.
This commit introduces an internal StateType to be able to differentiate
between the different types of state. This vastly improves performance.

Reviewed-by: Eskil Abrahamsen Blomfeldt