Fixes regression in QComboBox with keypad navigation while mouse is over the popup
The problem is that is we move the selection with the keyboard
in a way that causes the view to scroll, the current selection
would jump.

This happends since commit 9cb231d773db6deb that fixed the emission
of the enter signal when the view is scrolled.

We Therefor cannot use the enter signal.  Catch manually the mouse
move events to update the selection instead.

Reviewed-by: Thierry