Fix crash in QApplication autotest introduced by native pixmaps (S60 3.1)
Problem was caused by S60 closing a handle under our feet when QApplication
is destroyed. To avoid this, we open our own handle for the global lock
instead of using the shared one inside CFbsBitmap.
Also the global unlock/relock is not needed on S60 3.2 so this can be
eliminated and save a few cycles (the function call was a no-op before)

Reviewed-by: Jani Hautakangas