Fix for unresponsive sliders after orientation switch or full-screen video playback
During the switch to full-screen video playback, the following happens:

1. Double-tapping to enable full-screen results in a call to
QSymbianControl::HandleLongTapL
2. This modifies the global variable QApplication::mouse_buttons,
OR'ing in Qt::RightButton
3. QWidgetPrivate::create_sys, called as a result of the call to
setWindowFlags made by Phonon::VideoWIdget::setFullScreen, schedules a
delayed deletion of the same control as in step (1) above
4. The control gets deleted before it receives a HandlePointerEventL
for the long tap release, which would have removed Qt::RightButton
from the mouse_button bitmask
5. In subsequent calls to QSlider::mousePressEvent, the test
(ev->buttons() ^ ev->button()) is false, which results in the event
being ignored.

Ideally, we would fix this by propagating the m_previousEventLongPress
flag from the deleted QSymbianControl to the newly created one.
However, this does not work because the new control does not receive
the HandlePointerEventL callback for the long press release.

We therefore fix the bug by checking for m_previousEventLondPress
in the QSymbianControl destructor; if it is set, we clear the
Qt::RightButton bit from the QApplication::mouse_buttons mask.

Note that QTBUG-5309 (Cannot interact with sliders after orientation
switch during audio playback) is still seen after applying this patch.

Task-number: QTBUG-5242, QTBUG-5308
Reviewed-by: axis