Style's theme palette hash is lost after orientation switch
Currently, after orientation switch when new theme background is applied,
it is done using QApplication::setPalette(). This unfortunately tosses
away the palette hash. So instead directly calling
QApplication::setPalette(), the fix is to call the style's internal
setThemePalette(), which copies the hash also to QApplication palette.

This is related to bug QT-1478 in Private JIRA.

Task-number: QT-1478
Reviewed-by: Shane Kearns