Manual glyph shaping test.
It tests the glyph shaper in Qt by rendering a couple
of glyph combinations that are known to be strongly
shaped.
For now, this test is manual and simply generates
an html report for a visual test. It can, however also be
implemented as auto test.
The test data contains Vietnamese and Tamil data sets,
more will follow.

Reviewed-by: trustme