Call PositionChanged from QSymbianControl::SizeChanged
This is necessitated by the fact that CCoeControl::SetExtent calls
SizeChanged, but does not call PositionChanged.  This causes a bug
in the handling of orientation changes.  When the screen orientation
changes, QSymbianControl::HandleResourceChange(KInternalStatusPaneChange)
is called on the top-level widget.  This results in a call to
SetExtent.  Because PositionChanged was not being called, the outcome
was that the size of the widget's data.crect variable was updated, but
not its position.  This meant that mouse events sent to this widget
or any of its children were incorrectly handled:

1. QSymbianControl::HandlePointerEventL works out which widget should
receive the event. This uses CCoeControl::PositionRelativeToScreen,
and therefore calculates the correct widget.

2. In constructing the QMouseEvent object, QWidget::mapFromGlobal is
used to translate the event position into the receiving widget's
co-ordinate system. This uses data.crect.topLeft(), therefore giving
the wrong value, resulting in the widget's event handler discarding
the event.

Task-number: QTBUG-4697
Reviewed-by: axis