Cocoa: QColorDialog crashing when selection colors from system palette
The reason was that we did not handle the color space of the selected
color correctly. When selection from the system palette, the color
points to an index rather than e.g. RGB directly. This patch closes
the gap.

Another, a bit more evil, crash comes when trying to click on
'SelectedMenuItemColor' from the 'Developer' palette. The exact
same behaviour occurs when testing a native cocoa app in xcode
directly. So, to handle this as gracefully as possible, we sourround
the 'run modal' call with try-catch, and makes sure that we don't
quit the dialog until the user actually tells it to.

Bugreport to Apple created (bugreport.apple.com): 7364080

Task-number: QTBUG-4578
Reviewed-by: Prasanth