Fixed modal dialog fading support in Symbian.
Switched to use S60 API instead of Symbian one, the rationale is given
below as an copy/paste from S60 docs:

"Cross-application fading support

Symbian OS provides some support for fading, but this only covers
fading within an application. The S60 UI can have more than one
application on the screen at once (in fact, since the system parts
of the Status pane are inside EikServer, it always has more than one
application on the screen). The S60 UI introduced the fading and
drawing system to manage the fading state for the whole system.

Application and UI control code for S60 should not attempt to set fade
for a popup window directly. Instead, the popup window should implement
the MAknFadedComponent interface, and should use the TAknPopupFader
support class to set the fade.

The application UI base classes interact with the fading and drawing
system by informing it when the foreground status of the application
changes. They also use it to implement system wide fading."

There is also task QTBUG-5393 to implement fading support in future
without S60 dependency.

Task-number: QTBUG-5181
Reviewed-by: Sami Merila