Add min/pref/max size properties to QGraphicsWidget
QGraphicsLayouts and specially QGraphicsAnchorLayout makes large
use of min/pref/max sizes. Making it properties allow QtScript
bindings to take advantage of this and eases the use of Anchor
Layout on this cases.

The first use case for this patch was the use that Plasma makes
of it on it's javascript engine.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>