QPlainTextEdit redraw issue in QTextControl
There is an optimization in QTextControl, that it only requests
updates for single blocks if only a single block was changed. This
update was requested with the block's bounding rectangle.

In the case of disabled line wrapping in combination with a very long
line, a block change can result in a narrower block bounding rectangle.
This then resulted in pixel on the screen not being redrawn properly.

The fix enlarges the requested update rectangle horizontally.

The bug was discovered during inhouse Qt Creator testing.

Reviewed-by: Oswald Buddenhagen