Weather Anchor Layout example
This was a design that was not possible before with the other
layouts. To do this, the developer would need to hardcode all the
positions and resize would not work.

Besides the fact that we use "pixmap widgets" no it's not totally
usefull widgets, it shows how anchors can help developers an designers
with to create applications with new layouts.

Signed-off-by: Artur Duque de Souza <artur.souza@openbossa.org>