Ensure IDC and UIC3 dependencies are correct
After the refactoring to src/tools/tools.pro the dependencies for IDC
and UIC3 ended up being incorrect, due to the eval rules later in that
file. This patch adds IDC and UIC3 to SRC_SUBDIRS after Qt3Support, and
before ActiveQt, so the dependencies remain correct as before the
refactoring. The added condition of src/tools/tools.pro ensures that we
add the tools to the Makefile in src/tools, should we decide to generate
it, and compile from there.

Reviewed-by: Jason McDonald