QS60Style: Harmonize theme graphic with LineEdit and QTextEdit
In S60 the single line and multi line editors have same theme
background. We should share one as well. Now we are using fancy
notepad graphic with QTextEdit, but it makes QTextEdits to look
apart from QLineEdits and in some themes it is styled rather badly
(as it is not a central graphic item in a theme). Therefore it is
better to use one and same graphic for both widgets.

Task-number: QTBUG-5259
Reviewed-by: Alessandro Portale