Removed logging from Phonon MMF backend
This statement calls a function (RWindowBase::ClientHandle) which was
introduced in S60 3.2.  Although there is a runtime check, a build
made against 3.2 or above will fail on a 3.1 device.  This manifests
itself by the plugin failing to load.

The log statement is not really necessary anyway, because, for
window-owning controls, the window handle is the same value as the
CCoeControl* pointer.  This means that logging
RWindowBase::ClientHandle is redundant information.

Task-number: QTBUG-5406
Reviewed-by: trustme