Added support for streaming playback to Phonon MMF backend
Because the MIME type of the stream cannot always be deduced from the
URL, we assume that it is a video stream.  This is based on the
assumption that the video controllers will be capable of parsing
the container formats for audio-only, as well as video clips.  Note
that this assumption may not hold on all devices.

Note that most implementations of the MMF client APIs do not support
HTTP streaming (a.k.a. progressive download).  The backend has therefore
only been tested with RTSP streams - see the JIRA entry for further
details.

Task-number: QTBUG-4660
Reviewed-by: Frans Englich
(cherry picked from commit 3117e3a6a9c1bf95fc30ebee4d8d11b646cb7125)