Support composition mode source for drawing in dfb
Qt's/DirectFB's composition modes are not compatible for drawing
operations. We only support SourceOver and handle whether or not to
blend colors using DSDRAW_BLEND based on the alpha/opacity value.
If drawing with Source and an opaque color we can still handle the
operation using DirectFB. This has major impact since WebKit fills the
background of web pages using Source.

Reviewed-by: Donald Carr <donald.carr@nokia.com>