improve ordered message insertion
in a ts file, the top level categorization are the contexts, so it makes
no sense to make messages adjoint by file when the context is different.

a more clever algorithm which tries to order the entire contexts is
conceivable. that would be advantageous for files which use multiple
contexts subsequentially if each context appears in only one file.