Fixed loading of 8-bit grayscale TIFF files.
The PHOTOMETRIC_MINISBLACK and PHOTOMETRIC_MINISWHITE settings also
apply to 8-bit grayscale, so we need to check the bit depth of the image
as well. For convenience we also try to write the images using
PHOTOMETRIC_MINISBLACK and PHOTOMETRIC_MINISWHITE as opposed to
PHOTOMETRIC_PALETTE when we detect that the color table is grayscale.

Task-number: QTBUG-5459
Reviewed-by: Benjamin Poulain
(cherry picked from commit 845c69b2e780e1646f68ee99beda10768a1971e2)