Fixed bug where JPEG scaling resulted in an uninitialised pixel column.
Loading a JPEG image after calling QImageReader::setScaledSize() could
cause a 1 pixel wide column of uninitialised pixels at the far right of
the image. The bug was caused by rounding in the scaling algorithm.
Fixed by copying the last calculated pixel all the way to the end of
the row.

Task-number: QTBUG-3282
Reviewed-by: Trond