Autotest: prevent race-condition from causing a test failure.
The test requires receiving the NameOwnerChanged signal from the bus
before the voidSignal we emit ourselves. For whatever reason, the bus
apparently delays the emission of that signal, which could cause our
signal to be received first.

We should actually wait for the signal, but a simple processEvents is
enough in my tests.