Focus changes cause softkey area to flicker
Current implementation of softkeys causes flicker, since it re-draws
the softkeys for each event handled.

First re-draw initializes the softkeys to empty ones, and second one
draws the new updated softkeys onscreen. Due to initialization to
empty softkey command set, AVKON cannot deduce if the needed redraw
is necessary (AVKON does not redraw softkeys if the softkey commands
and/or texts won't change).

The fixed behaviour is that softkeys are set to -1 for each command
when initializing, but no call to SetCommandSetL is done. Now, AVKON
won't even try to draw the softkeys until we set the real command set
later.

Task-number: QTBUG-5436
Reviewed-by: Janne Anttila