QPlastiqueStyle: Fix painting QTreeView with gradient
When setting a gradian brush as a base for a QTreeWidget (with
stylesheet), the branch decoration would be plain black.
This is because the color() of a gradient brush is always black.

Fix it by using the base brush itself to paint the decoration if
the brush is not solid color.

Task-number: QTBUG-3816
Reviewed-by: Thierry