build lrelease bootstrapped
needed for build-time qm generation in qt itself.
the downsides are a) that the other bootstrapped tools grow by ~12kB
(on x86) due to qdatastream being pulled in by qbytearray and qstring
and b) that lrelease isn't l10n'd itself anymore (the latter could be
fixed by building a non-qobject qtranslator).